fix: Fixed app listening to IPv6 instead of IPv4
Due to a change in the way the app dependency works, not specifying
an address (even just '0.0.0.0') means that the program would listen to
an IPv6 address/port, instead of IPv4, which nearly everything is still
configured for. This meant that the client could not see the media
scanner. This commit fixes that, by making the app dependency listen
on an IPv4 address/port.