docs: Updated docs to have more Linux specific details
When working to get this software running on Linux, I had a few things
that tripped me up. This commit updates the docs with my findings, to
try and help make it easier for others to setup and use this software on
Linux, hopefully without as many obstacles in their way.