feat: rewrite TLS to return JSON and optional Graphics Data Definition
This PR rewrites the TLS command to return a single JSON object containing a `templates` array. It also implements support for Graphics Data Definitions as specified by https://github.com/SuperFlyTV/GraphicsDataDefinition

BREAKING CHANGE: The TLS response is now a JSON object with a `templates` array property. The items in this array are objects with two properties:

1. id (required) - a string containing the identifier (name) of the template as used by Caspar
2. gdd (optional) - The extracted GraphicsDataDefinition JSON, if any, from the template.

Example:
```
200 TLS OK {"templates":[{"id":"GFX/FULLSCREEN","gdd":{"title":"One-Line GFX Template","type":"object","properties":{"text0":{"description":"Text\r\ncontent","type":"string","maxLength":50,"gddType":"single-line","gddOptions":{}}}}},{"id":"GFX/HEAD"},{"id":"GFX/L3D"},{"id":"GFX/PREVIEWRENDERER"},{"id":"GFX/STRAP"},{"id":"GFX/TICKER"},{"id":"GFX/WIPE"}]}
```